afxdp/forwarding_build: relocate inline tests to forwarding_build_tests.rs
forwarding_build.rs is a flat file. 707-LOC inline test block extracted to sibling forwarding_build_tests.rs (531 LOC after rustfmt + 5-line header). Stub: `#[cfg(test)] #[path = "forwarding_build_tests.rs"] mod tests;`.

After: forwarding_build.rs drops from 1,478 → 953 LOC. Note: this file remains around the threshold because production is genuinely substantial — only 707 LOC of inline tests; production fns dominate.

cargo build --release: clean. cargo test --release: 878 passed.
